backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
openai==1.43.0
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab